♻️ Separate types and fix some type errors
