+++ src/app/(routes)/laboratory/human-interface-design-class/types.ts
@@ -1,0 +1,19 @@
+export type FoodItem = {
+	name: string;
+	local_name: string;
+	english_name: string;
+	description: string;
+};
+
+export type RestaurantItem = {
+	name: string;
+	address: string;
+	rating: number;
+	user_rating_count: number;
+	price_level: string;
+};
+
+export type Search = {
+	country: string;
+	city: string;
+};
